feat: build state trie from proofs in host instead
Moves the trie building process (from proofs) from the client to the
host. The client now only verifies the root before and after the state
transition. This saves a lot of cycles as the trie building process
is expensive.

As a result, the account/storage loading logic has also been rewritten,
since the proofs are no longer passed into the client. Instead, the
client receive "state requests" which contains keys (i.e. addresses/
slots), with the actual values read from the trie.

Also switches to use the MPT implementation from `zeth` in `rsp-mpt`.